Set up proper copy constructors for all classes
I was accidently doing
Vector2 temp = *this;
which was creating a pointer and modifying the original instead of
making a copy like it should.
Here is the fix Vector2 temp(*this);
which uses the new copy constructor
Also implemented Matrix3 determinant function and added new constructors
for creating a Vector or Matrix from a single float by setting all
values to that float.